% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,}
\alias{ANY-method}
\title{integrateIt package}
\usage{
integrateIt(f, n, int, approxtype)
}
\arguments{
\item{f}{is a function over some bounded interval \code{[a,b]} to some real value.}

\item{n}{is the number of times we want to partition the domain.}

\item{int}{is a vector that describes the interval of the domain.}

\item{approxtype}{is what the user selects to either use \code{'trapezoid'} or \code{'simpson'} method for approximation.}
}
\value{
object of one of the two classes--the same as the approximation methods. The return has three outputs:
\itemize{
\item{x} Subset of domain, partitioned
\item{y} Target of the function over the interval
\item{result} The numerical approximation
}
}
\description{
This package uses simpson's and trapezoid rule to produce integral approximations
}
\examples{

integrateIt(exp, 10, c(5,15), "simpson")

}
\seealso{
trapezoid-class, simpson-class
}
\author{
Zion Little: \email{l.zion@wustl.edu}
}
