% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson-class.R
\docType{class}
\name{simpson-class}
\alias{simpson-class}
\alias{initialize,}
\alias{integrateIt-method}
\alias{print,simpson-method}
\title{Simpsons rule output}
\usage{
\S4method{print}{simpson}(x)
}
\description{
\code{integrateIt} function creates objects of class \code{"simpson"}
}
\details{
Objects of this class have the slots
\itemize{
\item \code{result} Is the numerical approx of integral over interval \code{[a,b]} using simpson rule
\item \code{x} is a vector of values from partitions over \code{[a,b]}
\item \code{y} is the function f valued on the interval
}
}
\author{
Zion Little: \email{l.zion@wustl.edu}
}
