% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapezoid-class.R
\docType{class}
\name{trapezoid-class}
\alias{trapezoid-class}
\alias{initialize,}
\alias{integrateIt-method}
\alias{print,trapezoid-method}
\title{Trapezoid rule output}
\usage{
\S4method{print}{trapezoid}(x)
}
\description{
\code{integrateIt} function creates objects of class \code{"trapezoid"}
}
\details{
Objects of this class have the slots
\itemize{
\item \code{result} Is the numerical approx of integral over interval \code{[a,b]} using trapezoid rule
\item \code{x} is a vector of values from partitions over \code{[a,b]}
\item \code{y} is the function f valued on the interval
}
}
\author{
Zion Little: \email{l.zion@wustl.edu}
}
